make dpi setable

+++ src/WrkLst/DocxMustache/DocImage.php
@@ -38,13 +38,13 @@ class DocImage
         return false;
     }
 
-    public function ResampleImage($parent, $imgs, $k, $data)
+    public function ResampleImage($parent, $imgs, $k, $data, $dpi = 72)
     {
         \Storage::disk($parent->storageDisk)->put($parent->local_path.'word/media/'.$imgs[$k]['img_file_src'], $data);
 
         //rework img to new size and jpg format
         $img_rework = \Image::make($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_src']));
-        $img_rework2 = \Image::make($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_src']));
+        if($dpi!=72) $img_rework2 = \Image::make($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_src']));
 
         $imgWidth = $img_rework->width();
         $imgHeight = $img_rework->height();
@@ -80,21 +80,27 @@ class DocImage
         $new_height = $img_rework->height();
         $new_width = $img_rework->width();
 
-        //for storing the image in high dpi, so it has good quality on high dpi screens
-        $high_dpi = 160;
-        $img_rework2->resize(((int)$w*($high_dpi/72)), $h, function ($constraint) { // make high dpi version for actual storage
-            $constraint->aspectRatio();
-            $constraint->upsize();
-        });
-        $img_rework2->save($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
+        if($dpi!=72)
+        {
+            //for storing the image in high dpi, so it has good quality on high dpi screens
+            $img_rework2->resize(((int)$w*($dpi/72)), $h, function ($constraint) { // make high dpi version for actual storage
+                $constraint->aspectRatio();
+                $constraint->upsize();
+            });
+            $img_rework2->save($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
 
-        //set dpi of image to high dpi
-        $im = new \imagick();
-        $im->setResolution($high_dpi,$high_dpi);
-        $im->readImage($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
-        $im->writeImage($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
-        $im->clear();
-        $im->destroy();
+            //set dpi of image to high dpi
+            /*$im = new \imagick();
+            $im->setResolution($dpi,$dpi);
+            $im->readImage($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
+            $im->writeImage($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
+            $im->clear();
+            $im->destroy();
+            */
+        }
+        else {
+            $img_rework->save($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
+        }
 
         $parent->zipper->folder('word/media')->add($parent->StoragePath($parent->local_path.'word/media/'.$imgs[$k]['img_file_dest']));
